requirements.txt
appdirs==1.4.3
asgiref==3.2.10
boto3==1.17.57
botocore==1.20.57
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
distlib==0.3.0
distro==1.4.0
Django==3.1
django-cors-headers==3.7.0
django-crontab==0.7.1
django-filter==2.4.0
django-json==0.27
django-pandas==0.6.4
django-rest-swagger==2.2.0
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
gcloud==0.17.0
geographiclib==1.52
geopy==2.1.0
googleapis-common-protos==1.53.0
html5lib==1.0.1
httplib2==0.19.1
idna==2.8
importlib-metadata==4.5.0
ipaddr==2.2.0
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
jws==0.1.3
lockfile==0.12.2
Markdown==3.3.4
MarkupSafe==2.0.1
msgpack==0.6.2
numpy==1.19.5
oauth2client==3.0.0
openapi-codec==1.3.2
packaging==20.3
pandas==1.1.5
pep517==0.8.2
Pillow==8.1.2
progress==1.5
protobuf==3.18.0rc1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
pyfcm==1.5.1
PyJWT==1.7.1
pyparsing==2.4.6
Pyrebase==3.0.27
python-dateutil==2.8.1
python-jwt==2.0.1
pytoml==0.1.21
pytz==2021.1
requests==2.11.1
requests-toolbelt==0.7.0
retrying==1.3.3
rsa==4.7.2
s3transfer==0.4.2
simplejson==3.17.3
six==1.15.0
sqlparse==0.4.1
stripe==2.60.0
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
zipp==3.4.1
